+++ Projects/DesignSystem/Sources/bottomSheet/BottomPresentationController.swift
@@ -8,65 +8,6 @@
 
 import Foundation
 import UIKit
-import SnapKit
-
-public class YBBottomSheetViewController: UIViewController {
-    
-    fileprivate let sheetContentView = UIView()
-    fileprivate let indicatorView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .ybColor(.gray3)
-        return view
-    }()
-    public let containerView = UIView()
-    
-    public override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setViews()
-        view.addSubview(sheetContentView)
-        sheetContentView.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.trailing.equalToSuperview().offset(-24)
-            make.leading.equalToSuperview().offset(24)
-            make.bottom.equalToSuperview().offset(-50)
-        }
-        
-        sheetContentView.addSubview(indicatorView)
-        sheetContentView.addSubview(containerView)
-        indicatorView.snp.makeConstraints { make in
-            make.width.equalTo(60)
-            make.height.equalTo(4)
-            make.top.equalToSuperview().offset(14)
-            make.centerX.equalToSuperview()
-        }
-        
-        containerView.snp.makeConstraints { make in
-            make.top.equalTo(indicatorView).offset(28)
-            make.trailing.equalToSuperview().offset(-24)
-            make.leading.equalToSuperview().offset(24)
-            make.bottom.equalToSuperview().offset(-50)
-        }
-    }
-    
-    init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    public func setViews() {
-        view.backgroundColor = .clear
-        
-        sheetContentView.backgroundColor = .white
-        sheetContentView.layer.cornerRadius = 20
-    }
-}
-
-import Foundation
-import UIKit
 
 class BottomPresentationController: UIPresentationController {
